refactor(core): flatten logic/types into core/
With a single use case, splitting `core/logic/` and `core/types/` each
held one file and added nesting without payoff. Collocate the logic
beside its types (`window-merge.ts` + `window-merge.types.ts`) so the
pair is obvious at a glance. The split can be reintroduced if the
domain grows.
