+++ src/core/window-merge.types.ts
